project/requirement.txt
aioredis==1.3.0
amqp==2.5.1
apipkg==1.5
asgi-redis==1.0.0
asgiref==3.2.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
Authlib==0.13
autobahn==19.9.3
Automat==0.7.0
awscli==1.16.198
beautifulsoup4==4.8.1
billiard==3.6.0.0
boto3==1.9.188
botocore==1.12.188
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
CodeConvert==3.0.2
colorama==0.3.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
crontab==0.22.6
cryptography==2.7
cssutils==1.0.2
daphne==2.3.0
datadispatch==1.0.0
defusedxml==0.6.0
dj-static==0.0.6
Django==2.2.6
django-admin==2.0.0
django-braces==1.13.0
django-colorful==1.3
django-common-helpers==0.9.2
django-cors-headers==3.0.2
django-cron==0.5.1
django-crontab==0.7.1
django-elasticsearch-dsl==7.1.0
django-elasticsearch-dsl-drf==0.20.3
django-excel-response2==2.0.8
django-extensions==2.2.1
django-inlinecss==0.3.0
django-models-ext==1.1.9
django-nine==0.2.2
django-nose==1.4.6
django-oauth-toolkit==1.2.0
django-oauth2-provider==0.2.6.1
django-redis==4.10.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.1.2
django-security==0.11.3
django-short-url==1.1.4
django-shortcircuit==0.1.0
django-six==1.0.4
django-social-share==1.3.2
django-storages==1.7.1
djangorestframework==3.10.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.14
ecdsa==0.13.2
elastic-apm==5.3.0
elasticsearch==7.0.5
elasticsearch-dsl==7.1.0
entrypoints==0.3
execnet==1.7.1
Faker==0.9.1
Flask==1.1.1
furl==2.1.0
future==0.17.1
gevent==1.4.0
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
gunicorn==19.9.0
hiredis==1.0.0
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.18
incremental==17.5.0
ipaddress==1.0.22
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jeepney==0.4.1
Jinja2==2.10.1
jmespath==0.9.4
keyring==19.2.0
kombu==4.6.5
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.2
locustio==0.13.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==7.1.0
msgpack==0.6.1
msgpack-python==0.5.6
netifaces==0.10.4
nose==1.3.7
oauth==1.0.1
oauth2-provider==0.0
oauthlib==3.0.2
olefile==0.45.1
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==19.0
pbr==5.4.3
pexpect==4.2.1
pika==1.1.0
Pillow==6.1.0
pluggy==0.12.0
protobuf==3.0.0
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.7
pycparser==2.19
pyee==6.0.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymitter==0.2.3
PyNaCl==1.3.0
pynliner==0.8.0
pyparsing==2.4.2
pyRFC3339==1.1
pytest==5.2.2
pytest-cov==2.8.1
pytest-django==3.7.0
pytest-forked==1.1.3
pytest-xdist==1.30.0
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1
pyzmq==18.1.0
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.1
screen==1.0.1
SecretStorage==3.1.1
shortuuid==0.5.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
static3==0.7.0
tasks==2.5.0
testresources==2.0.1
text-unidecode==1.2
TimeConvert==1.5.0
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
tzlocal==2.0.0
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.25.6
uWSGI==2.0.18
vine==1.3.0
wadllib==1.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.6.0
